TD-5153 checks for existence of file before trying to delete to prevent error on cancel

+++ DigitalLearningSolutions.Web/Controllers/FrameworksController/ImportCompetencies.cs
@@ -267,7 +267,10 @@ namespace DigitalLearningSolutions.Web.Controllers.FrameworksController
         {
             var data = GetBulkUploadData();
             var frameworkId = data.FrameworkId;
-            FileHelper.DeleteFile(webHostEnvironment, data.CompetenciesFileName);
+            if (!string.IsNullOrWhiteSpace(data.CompetenciesFileName))
+            {
+                FileHelper.DeleteFile(webHostEnvironment, data.CompetenciesFileName);
+            }
             TempData.Clear();
             return RedirectToAction("ViewFramework", new { frameworkId, tabname = "Structure" });
         }
